+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # master
 
 ### unreleased
+- **fix:** refactor setup/teardown to avoid shared state
 
 ### [v0.6.1] - 2016-05-30
 - **fix:** rename closure parameters